requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.5.0
asgiref==3.5.0
cassandra-driver==3.25.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.1; python_version >= '3.6' and python_version < '4'
email-validator==1.1.3
fastapi==0.75.0
geomet==0.2.1.post1; python_version >= '2.7' and python_version != '3.3' and python_version < '4'
h11==0.13.0
idna==3.3
pydantic[email]==1.9.0
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
